size_t changed into uint_8

+++ shm/cargo.cpp
@@ -1,11 +1,11 @@
 #include "cargo.hpp"
 
-Cargo& Cargo::operator+=(size_t amount) {
+Cargo& Cargo::operator+=(uint8_t amount) {
     amount_ += amount;
     return *this;
 }
 
-Cargo& Cargo::operator-=(size_t amount) {
+Cargo& Cargo::operator-=(uint8_t amount) {
     amount_ -= amount;
     return *this;
 }
