+++ shm/cargo.hpp
@@ -4,11 +4,11 @@
 
 class Cargo {
    public:
-    Cargo& operator+=(size_t amount);
-    Cargo& operator-=(size_t amount);
+    Cargo& operator+=(uint8_t amount);
+    Cargo& operator-=(uint8_t amount);
 
    private:
     std::string name_;
-    size_t amount_;
-    size_t basePrice_;
+    uint8_t amount_;
+    uint8_t basePrice_;
 };
