+++ shm/ship.cpp
@@ -4,16 +4,16 @@
 
 Ship::Ship() : id_(-1){};
 
-Ship::Ship(int capacity, int maxCrew, int speed, const std::string& name, size_t id)
+Ship::Ship(uint8_t capacity, uint8_t maxCrew, uint8_t speed, const std::string& name, uint8_t id)
     : capacity_(capacity), maxCrew_(maxCrew), crew_(0), speed_(speed), name_(name), id_(id){};
 
-Ship::Ship(int maxCrew, int speed, size_t id) : Ship(0, maxCrew, speed, "", id){};
+Ship::Ship(uint8_t maxCrew, uint8_t speed, uint8_t id) : Ship(0, maxCrew, speed, "", id){};
 
 void Ship::setName(const std::string& name) {
     name_ = name;
 }
 
-Ship& Ship::operator-=(size_t crew) {
+Ship& Ship::operator-=(uint8_t crew) {
     if (crew_ < crew) {
         crew_ -= crew;
     }
@@ -21,7 +21,7 @@ Ship& Ship::operator-=(size_t crew) {
     return *this;
 }
 
-Ship& Ship::operator+=(size_t crew) {
+Ship& Ship::operator+=(uint8_t crew) {
     if (crew_ + crew <= maxCrew_) {
         crew_ += crew;
     }
@@ -29,7 +29,7 @@ Ship& Ship::operator+=(size_t crew) {
     return *this;
 }
 
-size_t Ship::getId() const {
+uint8_t Ship::getId() const {
     return id_;
 }
 
@@ -37,14 +37,14 @@ std::string Ship::getName() const {
     return name_;
 }
 
-size_t Ship::getSpeed() const {
+uint8_t Ship::getSpeed() const {
     return speed_;
 }
 
-size_t Ship::getMaxCrew() const {
+uint8_t Ship::getMaxCrew() const {
     return maxCrew_;
 }
 
-size_t Ship::getCapacity() const {
+uint8_t Ship::getCapacity() const {
     return capacity_;
 }
